fix: use int64 for block mask index arithmetic

+++ mlx/backend/cuda/gemms/block_mask.h
@@ -12,18 +12,18 @@ void apply_block_mask(
     array& data,
     const array& mask,
     int block_size,
-    int rows,
-    int cols,
+    int64_t rows,
+    int64_t cols,
     int64_t data_batch_stride,
-    int batch_count);
+    int64_t batch_count);
 
 array copy_with_block_mask(
     cu::CommandEncoder& encoder,
     const array& src,
     const array& mask,
     int block_size,
-    int rows,
-    int cols,
-    int batch_count);
+    int64_t rows,
+    int64_t cols,
+    int64_t batch_count);
 
 } // namespace mlx::core
